requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
crewai>=0.28.0
duckduckgo-search>=4.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0 